+++ src/main/java/com/batuhaniskr/project/repository/ProjectRepository.java
@@ -3,5 +3,5 @@ package com.batuhaniskr.project.repository;
 import com.batuhaniskr.project.model.Project;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProjectRepository extends JpaRepository<Project, Integer> {
+public interface ProjectRepository extends JpaRepository<Project, Long> {
 }
